chore: rename project to `sghi-etl-core` (#2)
After further RnD, it's now clear it is possible to have the same
abstractions for both simpler and complex ETL workflows. Complex
workflows in this context refer to ETL workflows that support
streaming and batching. Therefore, one API specifications project is
sufficient to handle both use cases. Hence, the renaming of the project
to a more generic name to emphasize that this is a general API for
describing ETL workflows.