chore(minor): deprecate `Processor.process` in favor of `apply` (#14)
Add a new method, `apply`, to the `sghi.etl.core.Processor` interface.
This will have the same role as the `process` method currently has on
the same interface. The verb _apply_ best describes how a `Processor`
is used. That is, you _draw_ raw data from a `Source`, _apply_ a
`Processor` to the drawn data to get clean data, and then you _drain_
the clean data to a `Sink`. Therefore, the `process` method now
delegates to the new `apply` method and is deprecated.